Rename image_buffer.{c,h} to image.{c,h}
