+++ src/compositor/buffer/image.h
